Add assembly.py
